+++ src/lib/components/ui/index.ts
@@ -171,6 +171,8 @@ export {
 	SheetDescription,
 	SheetFooter,
 	SheetHeader,
+	SheetOverlay,
+	SheetPortal,
 	SheetTitle,
 	SheetTrigger
 } from './sheet';